Use resilient function to get 3scale statistics
Tests are unstable because of delay within 3scale and it doesn't seem to
be smart idea to handle this in the client itself as this not generic
problem actually.

Therefore let's implement hardened methods/functions in the testsuite
(at least for now).

By this commit, new module to hold this is introduced and function to
get statistics it implemented.