Revert backoff on custom_account fixture
The backoff was supposed to mitigate 409 error returned right after the
deployment or custom tenant creation.

This didn't fix it as the API returned 422 on repeated calls and made
the situation even worse under some circumstances. The only efect was
that the faliure was significantly delayed and test execution prolonged.