Refactor test_email_accounts.py
It was bit difficult to understand the original code. While this
proposal might be hard to read as well for someone, I believe its flow
is much easier to follow and particular checks should be clearer.

Furthermore fixtures are now module-scoped, this is because of
backoff/flakiness. In this particular case it is not desirable to repeat
setup/cleanup each time test is executed, what ma be a risk in case of
function scoped fixtures.