Optimistic approving github action ;-)
This si a regular action that tries move forward stale PRs. If there is
a PR with 1 approve already that is 10 days old, then bot adds another
approval. This happens only if at least two reviews are requested. So
author has to request two reviews, one approval has to be granted and
that has to be 10 days old.

This doesn't mean automatic merge, it "just" unblocks the merge. It is
still on the merger if it is merged or not.