Deploy all tools to same namespace
tools target allows instalation of tools services in the openshift. This
change install all the tools to same shared namespaces. Some tools are
installed to 3scale specific namespace because they can't be easily
shared between 3scale instances. tools target is supposed to cover
deployments where just one 3scale is installed therefore it is not a
problem.