Detect and tag development release
Some tests should not be ran on dev release and they are causing false
alarms. There should be posibility to skip them, therefore it should be
possible to idenitfy dev release.

To avoid unwanted skips detection should be really conservative. For now
the implemetation determines it based on the name of the namespace. This
doesn't cover all possible cases when dev release is used, better to
have few false alarms than missed issue though.