Option to run all the tests in dedicated tenant
By default the tests are executed on default tenant or on that provided
by config.

Since now there is the option to run the tests on new/fresh tenant that
is created at the beginning of run and deleted at the end.

This covers the cases when custom tenant creation is covered by all the
tests or if it is easier to get master token than admin token (typically
true on deployments where a tenant for testing purposes is not present).

A caveat is that the testsuite creates as many tenant as many "threads"
are used to run it ('-n' option of xdist).