Update header accordingly in case of private tenant
admin url which is auto-detected is incorrect if private tenant is in
use. It is "difficult" (impossible) to know private tenant URL at the
moment of building the header (it would be possible if it was prepared
in dynaconf_loader, let's assume impossibility unless the absence of
correct tenant url is proven as mistake).