Backoff on 502 error in test_rate_limit_connection
For unknown reason api client returns 502 error in
test_rate_limit_connectyion.py right after expected 503 error.
Furthermore this is nearly impossible to reproduce & debug thus. So
here is a blind attempt to retry also on 502 error (that doesn't occur
in any other test however) as an attempt to eliminate this error.