Ability for custom cleanup handling of custom_service
A test with pretty large count of services (apicast pagination) uses
parallel service creation to safe time. There was no option for parallel
cleanup execution. This patch provides such option. It is not done by
default and must be handled individually in the test, however interface
is there. custom_service fixture is now object with unchanged behavior
that exposes list of finalizers of services that should not be
autocleaned. These finalizers then can and must be handled in particular
test(s).