Base class for 3scale custom buttons
Some 3scale buttons have patternfly class missing thefore customization
is needed. widgetastic has this class "hardcoded" in "_private" method
so customization of returned value is needed.

As a result this modified Button can be used in same way as original
widget, just on elements that do not contain relevant class.