Autouse threescale fixture
Some UI fixtures do not require threescale fixture. However it is a must
when private-tenant is used, because it modifies setup accordingly.
Therefore (at least for now) threescale fixture has to be autoused (it
is already session scoped) to ensure that the setup is done even for
fixtures/tests that do not have dependency on it.

There no or minimal trade-off. threescale fixture is "for free" in
normal setup, with private-tenant time and resources is needed to create
new tenant. However this is exactly desired purpose, all the tests
should run on private tenant in such case.